Add dummy words to stay comparable when needed
Documents might contain slightly out of sync token structures, when
artificial tokens were inserted in one but not the other.

To keep the program running, we enter something like an empty word,
which is effective, but can also be problematic in the final reports:
When a word is present only in the reviewable file, we cannot keep
track of it in the diff report.